moved phone number solution to external class
